BaseJob: further minor code cleanup

+++ jobs/basejob.h
@@ -168,7 +168,7 @@ namespace QMatrixClient
              * @param nextAttempt the 1-based number of attempt (will always be more than 1)
              * @param inMilliseconds the interval after which the next attempt will be taken
              */
-            void retryScheduled(size_t nextAttempt, int inMilliseconds);
+            void retryScheduled(int nextAttempt, int inMilliseconds);
 
             /**
              * Emitted when the job is finished, in any case. It is used to notify
@@ -185,7 +185,6 @@ namespace QMatrixClient
              * to avoid dangling pointers in your list.
              *
              * @param job the job that emitted this signal
-             * @internal
              *
              * @see success, failure
              */
